+++ yarn-project/aztec-node/package.json
@@ -3,7 +3,10 @@
   "version": "0.1.0",
   "main": "dest/index.js",
   "type": "module",
-  "exports": "./dest/index.js",
+  "exports": {
+    ".": "./dest/index.js",
+    "./config": "./dest/aztec-node/config.js"
+  },
   "bin": "./dest/bin/index.js",
   "typedocOptions": {
     "entryPoints": [
