+++ yarn-project/aztec/src/mnemonic.ts
@@ -1,0 +1,1 @@
+export const DefaultMnemonic = 'test test test test test test test test test test test junk';
