+++ yarn-project/circuit-types/src/config.ts
@@ -1,0 +1,2 @@
+export { ProverAgentConfig, proverAgentConfigMappings } from './interfaces/prover-agent.js';
+export { ProverBrokerConfig, proverBrokerConfigMappings } from './interfaces/prover-broker.js';
