+++ yarn-project/cli-wallet/src/cmds/check_tx.ts
@@ -1,5 +1,5 @@
 import { type PXE, type TxHash } from '@aztec/aztec.js';
-import { inspectTx } from '@aztec/cli/utils';
+import { inspectTx } from '@aztec/cli/inspect';
 import { type LogFn } from '@aztec/foundation/log';
 
 export async function checkTx(client: PXE, txHash: TxHash, statusOnly: boolean, log: LogFn) {
