+++ yarn-project/cli/package.json
@@ -10,7 +10,8 @@
     "./pxe": "./dest/cmds/pxe/index.js",
     "./cli-utils": "./dest/utils/index.js",
     "./misc": "./dest/cmds/misc/index.js",
-    "./utils": "./dest/utils/index.js"
+    "./utils": "./dest/utils/index.js",
+    "./inspect": "./dest/utils/inspect.js"
   },
   "typedocOptions": {
     "entryPoints": [
