+++ yarn-project/cli/src/utils/commands.ts
@@ -5,7 +5,7 @@ import { Fr } from '@aztec/aztec.js/fields';
 import { LogId } from '@aztec/aztec.js/log_id';
 import { TxHash } from '@aztec/aztec.js/tx_hash';
 import { type PXE } from '@aztec/circuit-types';
-import { PublicKeys } from '@aztec/circuits.js';
+import { PublicKeys } from '@aztec/circuits.js/types';
 import { type LogFn } from '@aztec/foundation/log';
 
 import { type Command, CommanderError, InvalidArgumentError, Option } from 'commander';
