+++ yarn-project/cli/src/utils/index.ts
@@ -2,4 +2,3 @@ export * from './commands.js';
 export * from './aztec.js';
 export * from './encoding.js';
 export * from './github.js';
-export * from './inspect.js';
