+++ yarn-project/proof-verifier/package.json
@@ -2,7 +2,8 @@
   "name": "@aztec/proof-verifier",
   "type": "module",
   "exports": {
-    ".": "./dest/index.js"
+    ".": "./dest/index.js",
+    "./config": "./dest/config.js"
   },
   "scripts": {
     "build": "yarn clean && tsc -b",
