+++ yarn-project/prover-node/package.json
@@ -3,7 +3,8 @@
   "version": "0.1.0",
   "type": "module",
   "exports": {
-    ".": "./dest/index.js"
+    ".": "./dest/index.js",
+    "./config": "./dest/config.js"
   },
   "inherits": [
     "../package.common.json"
